fix: extract good-news business logic from metrics route to service
P2-014: Move deriveGoodNews() to server/services/good-news.ts so it
can be unit tested independently and reused by other consumers.
